Overlays and Car tweaks
Learned how to use Unity Overlays.
Created an Overlay that displays the data of a Wheel when it is selected.
Created an Overlay that will be used as a space to put useful buttons or values. For now, has a button that reloads the current scene in Edit or Play mode for testing.