big level editor additions
Instead of trying to make an editor that works in play and edit mode, focusing on just edit mode version for now.
Points can now be added, deleted, moved, etc, in the scene using provided overlay or simply interacting with the Point object in the Hierarchy.
Duplicated the old version of the script with the input handling for the play mode version.