even MORE Debug Level Creator additions
Added comments to WheelInfoDisplay.cs.
Added methods to Debug LC so setup is entirely automated, such as adding the LevelPoint tag to the project and a button in the app Menu to create a Debug LC in the scene.
Points can be created either at the end of the list, inserted at the selection index, or inserted after the selected index.
Changed how the overlay is handled. It is added to the overlay list in onenable and removed in ondisable, letting the user toggle its visibility in that time.
Added Display Options as a separate class that holds values for drawn point size, drawn point colour, and point connection line colour. Will likely add more ways to customise.