Set up functionality for create terrain GUI
Added names to TerrainData, will be used for filename.
Unsure if I should use SOs to store the data or just save with binary, what is better?
Created a confirm back button that asks the user to confirm they want to go back before the edit mode is actually changed.
Changed the EditMode enum again, there should be an editmode for each menu in the overlay, as a new menu means new functionality which means a new mode.
Need to have the Terrain layer be added automatically by the Debug LC, like the tags.